+++ express/blocks/layouts/layouts.js
@@ -13,40 +13,10 @@
 import {
   createTag,
   getIcon,
-// eslint-disable-next-line import/no-unresolved
 } from '../../scripts/scripts.js';
-
-function masonrize(masonry, $masonry) {
-  const width = window.innerWidth > 1400 ? 1400 : window.innerWidth;
-  const numCols = Math.floor((width - 64) / 252);
-  if (numCols !== masonry.numCols) {
-    masonry.numCols = numCols;
-    const columns = [];
-    for (let i = 0; i < numCols; i += 1) {
-      columns.push({
-        cells: [],
-        outerHeight: 0,
-      });
-    }
-
-    masonry.cells.forEach((cell) => {
-      const minOuterHeight = Math.min(...columns.map((column) => column.outerHeight));
-      const column = columns.find((col) => col.outerHeight === minOuterHeight);
-      column.cells.push(cell);
-      column.outerHeight += cell.outerHeight;
-    });
-
-    $masonry.innerHTML = '';
-
-    columns.forEach((column) => {
-      const $column = createTag('div', { class: 'masonry-col' });
-      column.cells.forEach((cell) => {
-        $column.append(cell.element);
-      });
-      $masonry.append($column);
-    });
-  }
-}
+import {
+  Masonry,
+} from '../shared/masonry.js';
 
 export default function decorate($block) {
   const $layouts = Array.from($block.children);
@@ -88,14 +58,9 @@ export default function decorate($block) {
     $block.append($layout);
   });
 
-  const cells = [...$block.children].map(($cell) => ({
-    element: $cell,
-    outerHeight: $cell.offsetHeight,
-  }));
-
-  const masonry = { numCols: 0, cells };
-  masonrize(masonry, $block);
+  const masonry = new Masonry($block, [...$block.children]);
+  masonry.draw();
   window.addEventListener('resize', () => {
-    masonrize(masonry, $block);
+    masonry.draw();
   });
 }
